+++ packages/stream_chat/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Upcoming
+## 6.9.0
 
 🐞 Fixed